+++ nuxt.config.ts
@@ -12,9 +12,16 @@ export default defineNuxtConfig({
     '@nuxt/test-utils',
     '@pinia/nuxt'
   ],
+  
   ui: {
     icons: {
       provider: 'iconify', 
     },
   },
+
+  runtimeConfig: {
+    public: {
+      apiBase: 'http://127.0.0.1:8000/api/'
+    }
+  }
 })