+++ CHANGELOG.md
@@ -17,3 +17,8 @@
 - added Category.vue, Status.vue, Select.vue, and Date.vue in pages/components.vue
 - modified Button.vue Icon.vue in components/nav/
 - modified NavbarUser.vue SidebarAdmin.vue in components/layout/
+
+### V1.0.3
+- added useApi.ts for API base URL 
+- added auth.ts for fetching and guest middleware
+- added login and register pages